rename all packages in core applying kotlin style guide https://android.github.io/kotlin-guides/style.html#package-names
