+++ app/src/main/java/org/stepic/droid/core/downloadingprogress/DownloadingPresenter.kt
