+++ app/src/main/java/org/stepic/droid/core/downloadingprogress/DownloadingView.kt
