+++ app/src/main/java/org/stepic/droid/core/updatingstep/UpdatingStepPosterImpl.kt
@@ -1,8 +1,8 @@
-package org.stepic.droid.core.updating_step
+package org.stepic.droid.core.updatingstep
 
 import org.stepic.droid.base.ListenerContainer
-import org.stepic.droid.core.updating_step.contract.UpdatingStepListener
-import org.stepic.droid.core.updating_step.contract.UpdatingStepPoster
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepListener
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepPoster
 import javax.inject.Inject
 
 class UpdatingStepPosterImpl
