+++ app/src/main/java/org/stepic/droid/core/updatingstep/contract/UpdatingStepListener.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.core.updating_step.contract
+package org.stepic.droid.core.updatingstep.contract
 
 interface UpdatingStepListener {
     fun onNeedUpdate(stepId: Long, isSuccessAttempt: Boolean)
