+++ app/src/main/java/org/stepic/droid/core/videomoves/contract/VideosMovedListener.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.core.video_moves.contract
+package org.stepic.droid.core.videomoves.contract
 
 interface VideosMovedListener{
     fun onVideosMoved()
