+++ app/src/main/java/org/stepic/droid/di/AppCoreModule.kt
@@ -28,12 +28,12 @@ import org.stepic.droid.concurrency.SingleThreadExecutor
 import org.stepic.droid.configuration.Config
 import org.stepic.droid.configuration.ConfigImpl
 import org.stepic.droid.core.*
-import org.stepic.droid.core.internet_state.InternetEnabledPosterImpl
-import org.stepic.droid.core.internet_state.contract.InternetEnabledListener
-import org.stepic.droid.core.internet_state.contract.InternetEnabledPoster
-import org.stepic.droid.core.video_moves.VideosMovedPosterImpl
-import org.stepic.droid.core.video_moves.contract.VideosMovedListener
-import org.stepic.droid.core.video_moves.contract.VideosMovedPoster
+import org.stepic.droid.core.internetstate.InternetEnabledPosterImpl
+import org.stepic.droid.core.internetstate.contract.InternetEnabledListener
+import org.stepic.droid.core.internetstate.contract.InternetEnabledPoster
+import org.stepic.droid.core.videomoves.VideosMovedPosterImpl
+import org.stepic.droid.core.videomoves.contract.VideosMovedListener
+import org.stepic.droid.core.videomoves.contract.VideosMovedPoster
 import org.stepic.droid.di.qualifiers.BackgroundScheduler
 import org.stepic.droid.di.qualifiers.MainScheduler
 import org.stepic.droid.fonts.FontsProvider
