+++ app/src/main/java/org/stepic/droid/di/AppStepModule.kt
@@ -6,9 +6,9 @@ import org.stepic.droid.base.Client
 import org.stepic.droid.base.ClientImpl
 import org.stepic.droid.base.ListenerContainer
 import org.stepic.droid.base.ListenerContainerImpl
-import org.stepic.droid.core.updating_step.UpdatingStepPosterImpl
-import org.stepic.droid.core.updating_step.contract.UpdatingStepListener
-import org.stepic.droid.core.updating_step.contract.UpdatingStepPoster
+import org.stepic.droid.core.updatingstep.UpdatingStepPosterImpl
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepListener
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepPoster
 
 @Module
 interface AppStepModule {
