+++ app/src/main/java/org/stepic/droid/di/course/CourseAndSectionsModule.kt
@@ -2,8 +2,8 @@ package org.stepic.droid.di.course
 
 import dagger.Binds
 import dagger.Module
-import org.stepic.droid.core.downloadingProgress.ProgressWatcher
-import org.stepic.droid.core.downloadingProgress.SectionProgressWatcher
+import org.stepic.droid.core.downloadingprogress.ProgressWatcher
+import org.stepic.droid.core.downloadingprogress.SectionProgressWatcher
 
 @Module
 interface CourseAndSectionsModule {
