+++ app/src/main/java/org/stepic/droid/receivers/InternetConnectionEnabledReceiver.java
@@ -10,8 +10,8 @@ import org.stepic.droid.analytic.Analytic;
 import org.stepic.droid.base.App;
 import org.stepic.droid.concurrency.MainHandler;
 import org.stepic.droid.core.LocalProgressManager;
-import org.stepic.droid.core.internet_state.contract.InternetEnabledPoster;
-import org.stepic.droid.core.updating_step.contract.UpdatingStepPoster;
+import org.stepic.droid.core.internetstate.contract.InternetEnabledPoster;
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepPoster;
 import org.stepic.droid.model.Step;
 import org.stepic.droid.storage.StoreStateManager;
 import org.stepic.droid.storage.operations.DatabaseFacade;
