+++ app/src/main/java/org/stepic/droid/ui/adapters/SectionAdapter.java
@@ -23,7 +23,7 @@ import org.stepic.droid.R;
 import org.stepic.droid.analytic.Analytic;
 import org.stepic.droid.base.App;
 import org.stepic.droid.core.ScreenManager;
-import org.stepic.droid.core.downloadingProgress.DownloadingPresenter;
+import org.stepic.droid.core.downloadingprogress.DownloadingPresenter;
 import org.stepic.droid.core.presenters.CalendarPresenter;
 import org.stepic.droid.core.presenters.DownloadingInteractionPresenter;
 import org.stepic.droid.model.Course;
