+++ app/src/main/java/org/stepic/droid/ui/fragments/LessonFragment.java
@@ -33,7 +33,7 @@ import org.stepic.droid.core.presenters.LessonPresenter;
 import org.stepic.droid.core.presenters.StepsTrackingPresenter;
 import org.stepic.droid.core.presenters.contracts.LessonTrackingView;
 import org.stepic.droid.core.presenters.contracts.LessonView;
-import org.stepic.droid.core.updating_step.contract.UpdatingStepListener;
+import org.stepic.droid.core.updatingstep.contract.UpdatingStepListener;
 import org.stepic.droid.model.Lesson;
 import org.stepic.droid.model.PersistentLastStep;
 import org.stepic.droid.model.Section;
