+++ app/src/main/java/org/stepic/droid/ui/fragments/SectionsFragment.java
@@ -44,8 +44,8 @@ import org.stepic.droid.base.App;
 import org.stepic.droid.base.Client;
 import org.stepic.droid.base.FragmentBase;
 import org.stepic.droid.core.LocalProgressManager;
-import org.stepic.droid.core.downloadingProgress.DownloadingPresenter;
-import org.stepic.droid.core.downloadingProgress.DownloadingView;
+import org.stepic.droid.core.downloadingprogress.DownloadingPresenter;
+import org.stepic.droid.core.downloadingprogress.DownloadingView;
 import org.stepic.droid.core.dropping.contract.DroppingListener;
 import org.stepic.droid.core.presenters.CalendarPresenter;
 import org.stepic.droid.core.presenters.CourseFinderPresenter;
